Refactor TRAP logParser : Splitting functions into dedicated modules
a. split up the major/lengthy functions in logParser.py and kept them in separate files
b. parse assert info -> assertInfo.py
c. find crash point -> crashPoint.py
d. parse tcb info -> tcbinfo.py
e. cli/logAnalyser/log_parser : kept all these related files in this folder
f. also moved heapNode.py to this folder, as its only called from logParser.py
